+++ NEWS.md
@@ -2,6 +2,9 @@
 
 # Next
 
+* New built-in frontend using "real graphical" widget for the popup
+  ([#1525](https://github.com/company-mode/company-mode/pull/1525)).
+  This also adds a hard dependency on the package `posframe`.
 * The default light theme colors were changed to a more neutral set, and the
   scroll bar background was removed
   ([#1529](https://github.com/company-mode/company-mode/pull/1529)).